Restructured to make crate prefer neither sync or async.
