+++ src/result.rs
@@ -8,9 +8,9 @@ use std::fmt;
 use hyper::Error as HttpError;
 use url::ParseError;
 
-#[cfg(any(feature="ssl", feature="async-ssl"))]
+#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
 use native_tls::Error as TlsError;
-#[cfg(any(feature="ssl", feature="async-ssl"))]
+#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
 use native_tls::HandshakeError as TlsHandshakeError;
 
 /// The type used for WebSocket results
@@ -38,13 +38,13 @@ pub enum WebSocketError {
 	/// A WebSocket URL error
 	WebSocketUrlError(WSUrlErrorKind),
 	/// An SSL error
-	#[cfg(any(feature="ssl", feature="async-ssl"))]
+	#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
 	TlsError(TlsError),
 	/// an ssl handshake failure
-	#[cfg(any(feature="ssl", feature="async-ssl"))]
+	#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
 	TlsHandshakeFailure,
 	/// an ssl handshake interruption
-	#[cfg(any(feature="ssl", feature="async-ssl"))]
+	#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
 	TlsHandshakeInterruption,
 	/// A UTF-8 error
 	Utf8Error(Utf8Error),
@@ -69,11 +69,11 @@ impl Error for WebSocketError {
 			WebSocketError::IoError(_) => "I/O failure",
 			WebSocketError::HttpError(_) => "HTTP failure",
 			WebSocketError::UrlError(_) => "URL failure",
-			#[cfg(any(feature="ssl", feature="async-ssl"))]
+			#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
 			      WebSocketError::TlsError(_) => "TLS failure",
-			#[cfg(any(feature="ssl", feature="async-ssl"))]
+			#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
             WebSocketError::TlsHandshakeFailure => "TLS Handshake failure",
-			#[cfg(any(feature="ssl", feature="async-ssl"))]
+			#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
             WebSocketError::TlsHandshakeInterruption => "TLS Handshake interrupted",
 			WebSocketError::Utf8Error(_) => "UTF-8 failure",
 			WebSocketError::WebSocketUrlError(_) => "WebSocket URL failure",
@@ -85,7 +85,7 @@ impl Error for WebSocketError {
 			WebSocketError::IoError(ref error) => Some(error),
 			WebSocketError::HttpError(ref error) => Some(error),
 			WebSocketError::UrlError(ref error) => Some(error),
-			#[cfg(any(feature="ssl", feature="async-ssl"))]
+			#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
 			      WebSocketError::TlsError(ref error) => Some(error),
 			WebSocketError::Utf8Error(ref error) => Some(error),
 			WebSocketError::WebSocketUrlError(ref error) => Some(error),
@@ -115,14 +115,14 @@ impl From<ParseError> for WebSocketError {
 	}
 }
 
-#[cfg(any(feature="ssl", feature="async-ssl"))]
+#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
 impl From<TlsError> for WebSocketError {
 	fn from(err: TlsError) -> WebSocketError {
 		WebSocketError::TlsError(err)
 	}
 }
 
-#[cfg(any(feature="ssl", feature="async-ssl"))]
+#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
 impl<T> From<TlsHandshakeError<T>> for WebSocketError {
 	fn from(err: TlsHandshakeError<T>) -> WebSocketError {
 		match err {
