+++ src/server/mod.rs
@@ -1,13 +1,13 @@
 //! Provides an implementation of a WebSocket server
-#[cfg(feature="ssl")]
+#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
 use native_tls::TlsAcceptor;
-pub use self::upgrade::{Request, HyperIntoWsError};
 
 pub mod upgrade;
 
 #[cfg(feature="async")]
 pub mod async;
 
+#[cfg(feature="sync")]
 pub mod sync;
 
 /// Marker struct for a struct not being secure
@@ -19,7 +19,7 @@ pub struct NoTlsAcceptor;
 /// is running over SSL or not.
 pub trait OptionalTlsAcceptor {}
 impl OptionalTlsAcceptor for NoTlsAcceptor {}
-#[cfg(feature="ssl")]
+#[cfg(any(feature="sync-ssl", feature="async-ssl"))]
 impl OptionalTlsAcceptor for TlsAcceptor {}
 
 pub struct WsServer<S, L>
